Remove LocalAuthViewModel; migrate to Auth0
