+++ frontend/app/src/main/java/com/example/bookkeeper/ui/theme/Theme.kt
@@ -69,7 +69,10 @@ fun BookKeeperTheme(
     if (!view.isInEditMode) {
         SideEffect {
             val window = (view.context as Activity).window
+
+            @Suppress("DEPRECATION")
             window.statusBarColor = colorScheme.primary.toArgb()
+
             WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = !darkTheme
         }
     }
